test and fix bugs in happy flow

+++ src/main/java/com/example/demo/dto/PostResponse.java
@@ -14,7 +14,7 @@ public class PostResponse {
     private String post;
     private UserDto postedBy; //User Id who posted.
     private Integer likesCount;
-    private Long commentsCount;
+    private Integer commentsCount;
     private Date createdAt;
     private Date updatedAt;
